Split log row to 2 components
